Feature/misu

+++ src/main/java/Team_Mute/back_end/domain/reservation/dto/response/ReservationResponseDto.java
@@ -16,7 +16,7 @@ public class ReservationResponseDto {
 	private String spaceName;
 	private Long userId;
 	private String userName;
-	private Long reservationStatusId;
+	private Integer reservationStatusId;
 	private String reservationStatusName;
 	private Integer reservationHeadcount;
 	private LocalDateTime reservationFrom;
